+++ services/dbaas/mariadb/backups.go
@@ -3,24 +3,27 @@ package mariadb
 import (
 	"context"
 	"fmt"
+
 	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	mariadb "github.com/ionos-cloud/sdk-go-bundle/products/dbaas/mariadb/v2"
+	"github.com/ionos-cloud/sdk-go-bundle/shared"
 	"github.com/ionos-cloud/sdk-go-bundle/shared/fileconfiguration"
-	mariadb "github.com/ionos-cloud/sdk-go-dbaas-mariadb"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/loadedconfig"
 )
 
 // GetClusterBackups retrieves a list of backups for a given cluster ID and the location in which the cluster is created.
-func (c *Client) GetClusterBackups(ctx context.Context, clusterID, location string) (mariadb.BackupList, *mariadb.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+func (c *Client) GetClusterBackups(ctx context.Context, clusterID, location string) (mariadb.BackupList, *shared.APIResponse, error) {
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	backups, apiResponse, err := c.sdkClient.BackupsApi.ClusterBackupsGet(ctx, clusterID).Execute()
 	apiResponse.LogInfo()
 	return backups, apiResponse, err
 }
 
 // FindBackupByID retrieves a backup by its ID and the location in which the cluster is created.
-func (c *Client) FindBackupByID(ctx context.Context, backupID, location string) (mariadb.BackupResponse, *mariadb.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+func (c *Client) FindBackupByID(ctx context.Context, backupID, location string) (mariadb.BackupResponse, *shared.APIResponse, error) {
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	backups, apiResponse, err := c.sdkClient.BackupsApi.BackupsFindById(ctx, backupID).Execute()
 	apiResponse.LogInfo()
 	return backups, apiResponse, err
@@ -47,7 +50,7 @@ func SetMariaDBClusterBackupsData(d *schema.ResourceData, retrievedBackups []mar
 			backupEntry["size"] = *retrievedBackup.Properties.Size
 		}
 		var baseBackupsToBeSet []interface{}
-		for _, baseBackup := range *retrievedBackup.Properties.BaseBackups {
+		for _, baseBackup := range retrievedBackup.Properties.BaseBackups {
 			baseBackupEntry := make(map[string]interface{})
 			if baseBackup.Size != nil {
 				baseBackupEntry["size"] = *baseBackup.Size