+++ services/dbaas/mariadb/client.go
@@ -2,16 +2,15 @@ package mariadb
 
 import (
 	"fmt"
-	"log"
 	"net/http"
 	"os"
 	"runtime"
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/meta"
-	"github.com/ionos-cloud/sdk-go-bundle/shared"
+	mariadb "github.com/ionos-cloud/sdk-go-bundle/products/dbaas/mariadb/v2"
+	shared "github.com/ionos-cloud/sdk-go-bundle/shared"
 	"github.com/ionos-cloud/sdk-go-bundle/shared/fileconfiguration"
-	mariadb "github.com/ionos-cloud/sdk-go-dbaas-mariadb"
 
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/clientoptions"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
@@ -29,68 +28,37 @@ func (c *Client) GetFileConfig() *fileconfiguration.FileConfig {
 }
 
 // GetConfig returns the configuration of the client
-func (c *Client) GetConfig() *mariadb.Configuration {
+func (c *Client) GetConfig() *shared.Configuration {
 	return c.sdkClient.GetConfig()
 }
 
 func NewClient(clientOptions clientoptions.TerraformClientOptions, fileConfig *fileconfiguration.FileConfig) *Client {
-	newConfig := mariadb.NewConfiguration(clientOptions.Credentials.Username, clientOptions.Credentials.Password, clientOptions.Credentials.Token, clientOptions.Endpoint)
+	newConfig := shared.NewConfiguration(clientOptions.Credentials.Username, clientOptions.Credentials.Password, clientOptions.Credentials.Token, clientOptions.Endpoint)
 
-	if os.Getenv(constant.IonosDebug) != "" {
-		newConfig.Debug = true
-	}
 	newConfig.MaxRetries = constant.MaxRetries
 	newConfig.MaxWaitTime = constant.MaxWaitTime
 
-	newConfig.HTTPClient = &http.Client{Transport: utils.CreateTransport(clientOptions.SkipTLSVerify)}
 	newConfig.UserAgent = fmt.Sprintf(
 		"terraform-provider/%s_ionos-cloud-sdk-go-dbaas-mariadb/%s_hashicorp-terraform/%s_terraform-plugin-sdk/%s_os/%s_arch/%s",
 		clientOptions.Version, mariadb.Version, clientOptions.TerraformVersion,
 		meta.SDKVersionString(), runtime.GOOS, runtime.GOARCH, //nolint:staticcheck
 	)
 
-	return &Client{
+	client := &Client{
 		sdkClient:  mariadb.NewAPIClient(newConfig),
 		fileConfig: fileConfig,
 	}
+
+	client.sdkClient.GetConfig().HTTPClient = &http.Client{Transport: utils.CreateTransport(clientOptions.SkipTLSVerify)}
+	return client
 }
 
-// overrideClientEndpoint todo - after move to bundle, replace with generic function from fileConfig
-func (c *Client) overrideClientEndpoint(productName, location string) {
-	// whatever is set, at the end we need to check if the IONOS_API_URL_productname is set and use override the endpoint if yes
-	defer c.changeConfigURL(location)
-	if os.Getenv(mariadb.IonosApiUrlEnvVar) != "" {
-		fmt.Printf("[DEBUG] Using custom endpoint %s\n", os.Getenv(mariadb.IonosApiUrlEnvVar))
-		return
-	}
-	fileConfig := c.GetFileConfig()
-	if fileConfig == nil {
-		return
-	}
-	config := c.GetConfig()
-	if config == nil {
-		return
-	}
-	endpoint := fileConfig.GetProductLocationOverrides(productName, location)
-	if endpoint == nil {
-		log.Printf("[WARN] Missing endpoint for %s in location %s", productName, location)
-		return
-	}
-	config.Servers = mariadb.ServerConfigurations{
-		{
-			URL:         endpoint.Name,
-			Description: shared.EndpointOverridden + location,
-		},
-	}
-	config.HTTPClient.Transport = shared.CreateTransport(endpoint.SkipTLSVerify, endpoint.CertificateAuthData)
-}
-
-// changeConfigURL modifies the URL inside the client configuration.
+// ChangeConfigURL modifies the URL inside the client configuration.
 // This function is required in order to make requests to different endpoints based on location.
-func (c *Client) changeConfigURL(location string) {
+func (c *Client) ChangeConfigURL(location string) {
 	clientConfig := c.sdkClient.GetConfig()
 	if location == "" && os.Getenv(ionosAPIURLMariaDB) != "" {
-		clientConfig.Servers = mariadb.ServerConfigurations{
+		clientConfig.Servers = shared.ServerConfigurations{
 			{
 				URL: utils.CleanURL(os.Getenv(ionosAPIURLMariaDB)),
 			},
@@ -99,7 +67,7 @@ func (c *Client) changeConfigURL(location string) {
 	}
 	for _, server := range clientConfig.Servers {
 		if strings.EqualFold(server.Description, shared.EndpointOverridden+location) || strings.EqualFold(server.URL, locationToURL[location]) {
-			clientConfig.Servers = mariadb.ServerConfigurations{
+			clientConfig.Servers = shared.ServerConfigurations{
 				{
 					URL:         server.URL,
 					Description: shared.EndpointOverridden + location,
@@ -108,6 +76,7 @@ func (c *Client) changeConfigURL(location string) {
 			return
 		}
 	}
+
 }
 
 var (